The shop app updated.

+++ the-shop/screens/user/UserProductsScreen.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { FlatList, Platform, Button } from 'react-native';
+import { FlatList, Platform, Button, Alert } from 'react-native';
 import { useSelector, useDispatch } from 'react-redux';
 import { HeaderButtons, Item } from 'react-navigation-header-buttons';
 
@@ -15,7 +15,12 @@ const UserProductsScreen = props => {
     const products = useSelector(state => state.products.userProducts);
         
     const deleteProduct = prodId => {
-        dispatch(productsActions.deleteProduct(prodId));
+        Alert.alert('Are you sure?', 'Do you really want to delete this item?', [
+            { text: 'No!', style: 'default' },
+            { text: 'Yes!', style: 'destructive', onPress: () => {
+                dispatch(productsActions.deleteProduct(prodId));
+            } }
+        ]);
     };
 
     const redirectToEditProduct = prodId => {
